Renamed AboutDialog files to match the class name
