+++ qtEditor/AboutDialog.cpp
@@ -1,4 +1,4 @@
-#include "aboutdialog.h"
+#include "AboutDialog.h"
 #include "ui_aboutdialog.h"
 
 AboutDialog::AboutDialog(QWidget* parent) : QDialog(parent), ui(new Ui::AboutDialog) { ui->setupUi(this); }
