+++ qtEditor/AboutDialog.h
